+++ src/app/pages/user-forms/madrasareq-list/madrasareq-list.component.ts
@@ -4,6 +4,7 @@ import { Router } from '@angular/router';
 import { Madrasa } from 'src/app/madrasa';
 import { LoginServiceService } from 'src/app/services/login-service.service';
 import { MadrasaServiceService } from 'src/app/services/madrasa-service.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-madrasareq-list',
@@ -38,10 +39,30 @@ constructor(private madrasaSrve:MadrasaServiceService,private router:Router,publ
   
   }
   deleteMadrasa(id:number){
-  this.madrasaSrve.deleteMadrasa(id).subscribe(data=>{
-     console.log(data);
-    this.getAllMadrasa();
-  });
+    Swal.fire({
+      title:'Are you want to delete?',
+      text:'You will not able to recover it',
+      icon:'warning',
+      showCancelButton:true,
+      confirmButtonText:'Yes delete it',
+      cancelButtonText:'Not keep it',
+    }).then((result)=>{
+      if(result.value){
+        this.madrasaSrve.deleteMadrasa(id).subscribe(data=>{
+          console.log(data);
+         this.getAllMadrasa();
+       });
+
+        Swal.fire(
+          'deleted!',
+          'Application has been deleted.',
+          'success'
+        );
+      }else if(result.dismiss==Swal.DismissReason.cancel){
+        Swal.fire('Cancelled','Your Application is safe:)','error');
+      }
+    })
+
   }
   acceptApplication(id: number) {
     this.madrasaSrve.acceptApplication(id).subscribe(
@@ -76,6 +97,19 @@ constructor(private madrasaSrve:MadrasaServiceService,private router:Router,publ
     this.madrasaSrve.getAcceptedList().subscribe(data=>{
       this.madrast= data;
     })
+  }
+
+  getStatusStyle(status:any): {[key:string]:string}{
+    switch (status){
+      case 'Accepted':
+        return {'color':'green'};
+      case 'Rejected':
+        return {'color':'red'};
+      case 'Pending':
+        return {'color':'blue'};
+      default:
+        return {};      
+    }
   }
   }
